requirements.txt
Django==2.0.5
djangorestframework==3.8.2
factory-boy==2.11.1
